ZZALOG: Bug: remove compiler warnings. Bug: Dupe check progress not handled properly. Enhancement: Add RESTORE button to header editor. Bug: Exception when close header editor with window X button. Enhancement: Report prefix exception as a LOG: event, improve message. Bug: Remove Deleted marker as it confuses ordering for DXCC entity codes. Performance: Stop updating file viewer when using process bar.

+++ zzalog/band_view.cpp
@@ -94,7 +94,7 @@ bool band_view::load_data() {
 	streampos startpos = file.tellg();
 	file.seekg(0, ios::end);
 	streampos currpos = file.tellg();
-	int length = currpos - startpos;
+	int length = (int)(currpos - startpos);
 	// reposition back to beginning
 	file.seekg(0, ios::beg);
 	// Initialise progress
